refactor: move ProofTarget to reth-trie-common
Move the Target proof type from reth-trie::proof_v2 to reth-trie-common
as ProofTarget, making it available to lower-level crates. The struct
fields are now pub since they were already accessed via pub(crate).

A type alias Target = ProofTarget remains in proof_v2 for compatibility.
The sub_trie_prefix helper stays in proof_v2 as a free function since it
depends on increment_and_strip_trailing_zeros from that module.

Amp-Thread-ID: https://ampcode.com/threads/T-019c9475-4446-7287-9cf3-a147508f942f
