+++ crates/trie/trie/src/proof_v2/target.rs
@@ -1,70 +1,32 @@
 use crate::proof_v2::increment_and_strip_trailing_zeros;
-use alloy_primitives::B256;
 use reth_trie_common::Nibbles;
 
-/// Target describes a proof target. For every proof target given, the
-/// [`crate::proof_v2::ProofCalculator`] will calculate and return all nodes whose path is a prefix
-/// of the target's `key`.
-#[derive(Debug, Copy, Clone)]
-pub struct Target {
-    pub(crate) key: Nibbles,
-    pub(crate) min_len: u8,
-}
+/// Type alias for backward compatibility.
+pub type Target = reth_trie_common::ProofTarget;
 
-impl Target {
-    /// Returns a new [`Target`] which matches all trie nodes whose path is a prefix of this key.
-    pub fn new(key: B256) -> Self {
-        // SAFETY: key is a B256 and so is exactly 32-bytes.
-        let key = unsafe { Nibbles::unpack_unchecked(key.as_slice()) };
-        Self { key, min_len: 0 }
-    }
-
-    /// Returns the key the target was initialized with.
-    pub fn key(&self) -> B256 {
-        B256::from_slice(&self.key.pack())
-    }
-
-    /// Only match trie nodes whose path is at least this long.
-    ///
-    /// # Panics
-    ///
-    /// This method panics if `min_len` is greater than 64.
-    pub fn with_min_len(mut self, min_len: u8) -> Self {
-        debug_assert!(min_len <= 64);
-        self.min_len = min_len;
-        self
-    }
-
-    // A helper function for getting the largest prefix of the sub-trie which contains a particular
-    // target, based on its `min_len`.
-    //
-    // A target will only match nodes which share the target's prefix, where the target's prefix is
-    // the first `min_len` nibbles of its key. E.g. a target with `key` 0xabcd and `min_len` 2 will
-    // only match nodes with prefix 0xab.
-    //
-    // In general the target will only match within the sub-trie whose prefix is identical to the
-    // target's. However there is an exception:
-    //
-    // Given a trie with a node at 0xabc, there must be a branch at 0xab. A target with prefix 0xabc
-    // needs to match that node, but the branch at 0xab must be constructed order to know the node
-    // is at that path. Therefore the sub-trie prefix is the target prefix with a nibble truncated.
-    //
-    // For a target with an empty prefix (`min_len` of 0) we still use an empty sub-trie prefix;
-    // this will still construct the branch at the root node (if there is one). Targets with
-    // `min_len` of both 0 and 1 will therefore construct the root node, but only those with
-    // `min_len` of 0 will retain it.
-    #[inline]
-    fn sub_trie_prefix(&self) -> Nibbles {
-        let mut sub_trie_prefix = self.key;
-        sub_trie_prefix.truncate(self.min_len.saturating_sub(1) as usize);
-        sub_trie_prefix
-    }
-}
-
-impl From<B256> for Target {
-    fn from(key: B256) -> Self {
-        Self::new(key)
-    }
+// A helper function for getting the largest prefix of the sub-trie which contains a particular
+// target, based on its `min_len`.
+//
+// A target will only match nodes which share the target's prefix, where the target's prefix is
+// the first `min_len` nibbles of its key. E.g. a target with `key` 0xabcd and `min_len` 2 will
+// only match nodes with prefix 0xab.
+//
+// In general the target will only match within the sub-trie whose prefix is identical to the
+// target's. However there is an exception:
+//
+// Given a trie with a node at 0xabc, there must be a branch at 0xab. A target with prefix 0xabc
+// needs to match that node, but the branch at 0xab must be constructed order to know the node
+// is at that path. Therefore the sub-trie prefix is the target prefix with a nibble truncated.
+//
+// For a target with an empty prefix (`min_len` of 0) we still use an empty sub-trie prefix;
+// this will still construct the branch at the root node (if there is one). Targets with
+// `min_len` of both 0 and 1 will therefore construct the root node, but only those with
+// `min_len` of 0 will retain it.
+#[inline]
+pub(crate) fn sub_trie_prefix(target: &Target) -> Nibbles {
+    let mut sub_trie_prefix = target.key;
+    sub_trie_prefix.truncate(target.min_len.saturating_sub(1) as usize);
+    sub_trie_prefix
 }
 
 // A helper function which returns the first path following a sub-trie in lexicographical order.
@@ -105,7 +67,7 @@ pub(crate) fn iter_sub_trie_targets(
     // where the sub-trie prefixes are equal (to differentiate targets which match the root node and
     // those which don't).
     targets.sort_unstable_by(|a, b| {
-        a.sub_trie_prefix().cmp(&b.sub_trie_prefix()).then_with(|| a.min_len.cmp(&b.min_len))
+        sub_trie_prefix(a).cmp(&sub_trie_prefix(b)).then_with(|| a.min_len.cmp(&b.min_len))
     });
 
     // We now chunk targets, such that each chunk contains all targets belonging to the same
@@ -126,13 +88,13 @@ pub(crate) fn iter_sub_trie_targets(
     //          - [0x2, 0x2fa] w/ upper bound 0x3
     //          - [0x4c 0x4ce] w/ upper bound 0x4d
     //          - [0x4e]       w/ upper bound 0x4f
-    let mut upper_bound = targets.first().and_then(|t| sub_trie_upper_bound(&t.sub_trie_prefix()));
+    let mut upper_bound = targets.first().and_then(|t| sub_trie_upper_bound(&sub_trie_prefix(t)));
     let target_chunks = targets.chunk_by_mut(move |_, next| {
         if let Some(some_upper_bound) = upper_bound {
-            let sub_trie_prefix = next.sub_trie_prefix();
-            let same_chunk = sub_trie_prefix < some_upper_bound;
+            let prefix = sub_trie_prefix(next);
+            let same_chunk = prefix < some_upper_bound;
             if !same_chunk {
-                upper_bound = sub_trie_upper_bound(&sub_trie_prefix);
+                upper_bound = sub_trie_upper_bound(&prefix);
             }
             same_chunk
         } else {
@@ -143,7 +105,7 @@ pub(crate) fn iter_sub_trie_targets(
     // Map the chunks to the return type. Within each chunk we want targets to be sorted by their
     // key, as that will be the order they are checked by the `ProofCalculator`.
     target_chunks.map(move |targets| {
-        let prefix = targets[0].sub_trie_prefix();
+        let prefix = sub_trie_prefix(&targets[0]);
         let retain_root = targets[0].min_len == 0;
         targets.sort_unstable_by_key(|target| target.key);
         SubTrieTargets { prefix, targets, retain_root }
@@ -153,6 +115,7 @@ pub(crate) fn iter_sub_trie_targets(
 #[cfg(test)]
 mod tests {
     use super::*;
+    use alloy_primitives::B256;
 
     #[test]
     fn test_iter_sub_trie_targets() {
